# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "mem9 Server API",
    "version": "v1alpha2",
    "description": "OpenAPI specification for the mem9 Go server routes registered by server/internal/handler/handler.go."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "https://api.mem9.ai",
      "description": "Production API server"
    }
  ],
...
  "paths": {
...
    "/v1alpha1/mem9s": {
      "post": {
        "tags": [
          "Provisioning"
        ],
        "operationId": "provisionMem9s",
        "summary": "Provision a mem9 tenant",
        "description": "Creates a new tenant. The endpoint has no request body. Optional UTM query parameters are captured when present.",
        "parameters": [
          {
            "$ref": "#/components/parameters/UtmSource"
          },
          {
            "$ref": "#/components/parameters/UtmMedium"
          },
          {
            "$ref": "#/components/parameters/UtmCampaign"
          },
          {
            "$ref": "#/components/parameters/UtmContent"
          }
        ],
        "responses": {
          "201": {
            "description": "Tenant provisioned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProvisionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/v1alpha2/status": {
      "get": {
        "tags": [
          "Status"
        ],
        "operationId": "getKeyStatus",
        "summary": "Get API key status",
        "description": "Validates X-API-Key against the control-plane state and returns whether the key is active or inactive.",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "API key status.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key used by v1alpha2 routes."
      }
    },
    "parameters": {
...
      "UtmSource": {
        "name": "utm_source",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "UtmMedium": {
        "name": "utm_medium",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "UtmCampaign": {
        "name": "utm_campaign",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "UtmContent": {
        "name": "utm_content",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "responses": {
...
      "BadRequest": {
        "description": "Invalid request. For v1alpha2 API-key routes this also includes missing, invalid, or inactive X-API-Key values returned by the authentication middleware.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid authentication credentials.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Version, duplicate-key, or resource conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "Write conflict or temporarily unavailable backend.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ProvisionResponse": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "description": "Provisioned tenant ID."
          }
        }
      },
      "KeyStatusResponse": {
        "type": "object",
        "required": [
          "status"
        ],
        "properties": {
          "status": {
            "$ref": "#/components/schemas/KeyStatus"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "KeyStatus": {
        "type": "string",
        "enum": [
          "active",
          "inactive"
        ]
      },
...
    }
  }
}